requirements.txt
# -----------------------------
# Core API & Server
# -----------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9

# -----------------------------
# LLM & Agent Framework
# -----------------------------
openai>=1.12.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.6
langgraph>=0.0.45
crewai>=0.35.0
crewai-tools>=0.4.6

# -----------------------------
# Embeddings & Vector Store
# -----------------------------
chromadb>=0.4.24
tiktoken>=0.6.0

# -----------------------------
# Literature & Web Search
# -----------------------------
requests>=2.31.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3

# -----------------------------
# Data Handling & Statistics
# -----------------------------
pandas>=2.2.0
numpy>=1.26.4
scipy>=1.12.0
statsmodels>=0.14.1

# -----------------------------
# Visualization
# -----------------------------
matplotlib>=3.8.3
seaborn>=0.13.2

# -----------------------------
# Machine Learning
# -----------------------------
scikit-learn>=1.4.1

# -----------------------------
# Document Generation
# -----------------------------
reportlab>=4.1.0
python-docx>=1.1.0
openpyxl>=3.1.2
xlrd>=2.0.1

# -----------------------------
# Frontend (optional but recommended)
# -----------------------------
streamlit>=1.32.0

# -----------------------------
# Utilities
# -----------------------------
python-dotenv>=1.0.1
tqdm>=4.66.2
